--- .prettierignore
@@ -1,16 +1,0 @@
-# Jest files
-cache
-coverage
-
-# Expo files
-.expo
-
-# NPM files
-node_modules
-
-# dotenv environment variables file
-.env
-.env.test
-
-# Not Text
-assets
