--- .prettierrc.json
@@ -1,9 +1,0 @@
-{
-  "tabWidth": 2,
-  "singleQuote": true,
-  "bracketSpacing": true,
-  "printWidth": 80,
-  "semi": true,
-  "bracketSameLine": false,
-  "arrowParens": "always"
-}
