--- __tests__/App.test.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-// import { cleanup } from '@testing-library/react-native'
-// import { StyleSheet, Text, View } from 'react-native';
-
-// afterEach(cleanup);
-
-import App from '../App';
-
-// jest.mock('react-native', () => {
-//   return {
-//     StyleSheet: {
-//       create: {() => ({})},
-//     },
-//   };
-// });
-
-// jest.mock('react-native', () => ({
-//   StyleSheet: {
-//     create: function () { return "" }
-//   }
-// }))
-
-describe('<App />', () => {
-  // it('has 2 children', () => {
-  //   const tree = renderer.create(<App />).toJSON();
-  //   expect(tree.children.length).toBe(2);
-  // });
-
-  it('renders correctly', () => {
-    const tree = renderer.create(<App />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
