--- app.json
@@ -1,26 +1,0 @@
-{
-  "expo": {
-    "name": "Snomes",
-    "slug": "Snomes",
-    "version": "1.0.0",
-    "orientation": "portrait",
-    "splash": {
-      "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
-    },
-    "updates": {
-      "fallbackToCacheTimeout": 0
-    },
-    "assetBundlePatterns": ["**/*"],
-    "ios": {
-      "supportsTablet": true
-    },
-    "android": {
-      "adaptiveIcon": {
-        "backgroundColor": "#FFFFFF"
-      }
-    },
-    "web": {
-    }
-  }
-}
