--- cypress/integration/spec.js
@@ -1,6 +1,0 @@
-/// <reference types="cypress" />
-describe('page', () => {
-  it('works', () => {
-    cy.visit('https://example.cypress.io');
-  });
-});
