+++ frontend/.prettierignore
@@ -11,6 +11,3 @@ node_modules
 # dotenv environment variables file
 .env
 .env.test
-
-# Not Text
-assets
