+++ frontend/package.json
@@ -24,17 +24,12 @@
   },
   "dependencies": {
     "@babel/preset-env": "^7.16.0",
-    "@mantine/hooks": "^3.1.7",
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "axios": "^0.24.0",
     "expo": "~43.0.0",
     "expo-status-bar": "~1.1.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.2",
-    "react-native-material-textfield": "^0.16.1",
-    "react-native-web": "0.17.1",
-    "styled-components": "^5.3.3"
+    "react-native-web": "0.17.1"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
