--- frontend/src/navigation/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-
-import { AppNavigator } from './AppNavigator';
-
-export const Navigation = () => {
-
-    return (
-        <NavigationContainer>
-            <AppNavigator />
-        </NavigationContainer>
-    );
-};
