--- frontend/src/screens/MatchScreen.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { View, Text } from 'react-native'
-
-const MatchScreen = () => {
-    return (
-        <View>
-            <Text>Matches</Text>
-        </View>
-    )
-}
-
-export default MatchScreen
