--- package.json
@@ -1,62 +1,0 @@
-{
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject",
-    "test": "jest --watchAll --no-cache",
-    "cypress": "sudo docker run -it -v $PWD:/e2e -w /e2e cypress/included:9.0.0",
-    "cy:watch": "docker run -it -v $PWD:/e2e -w /e2e --entrypoint=cypress/included:9.0.0 info",
-    "prettier": "prettier --write ."
-  },
-  "jest": {
-    "preset": "jest-expo",
-    "transform": {
-      "^.+\\.[jt]sx?$": "babel-jest"
-    },
-    "testPathIgnorePatterns": [
-      "./cypress"
-    ],
-    "transformIgnorePatterns": [
-      "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)"
-    ],
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "**/*.{js,jsx}",
-      "!**/coverage/**",
-      "!**/node_modules/**",
-      "!**/babel.config.js",
-      "!**/jest.setup.js"
-    ]
-  },
-  "dependencies": {
-    "@babel/preset-env": "^7.16.0",
-    "@mantine/hooks": "^3.1.7",
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "@react-navigation/native": "^6.0.6",
-    "axios": "^0.24.0",
-    "babel-jest": "^27.3.1",
-    "expo": "~43.0.0",
-    "expo-status-bar": "~1.1.0",
-    "jest": "^26.6.3",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-native": "0.64.2",
-    "react-native-material-textfield": "^0.16.1",
-    "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "^3.8.0",
-    "react-native-web": "0.17.1",
-    "styled-components": "^5.3.3"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@testing-library/react-native": "^8.0.0",
-    "babel-preset-react-native": "^4.0.1",
-    "jest-expo": "^43.0.1",
-    "prettier": "2.4.1",
-    "react-test-renderer": "^17.0.2"
-  },
-  "private": true
-}
