--- src/screens/MessageScreen.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { View, Text } from 'react-native'
-
-const MessageScreen = () => {
-    return (
-        <View>
-            <Text>Messages</Text>
-        </View>
-    )
-}
-
-export default MessageScreen
