--- src/screens/ProfileScreen.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { View, Text } from 'react-native'
-
-const ProfileScreen = () => {
-    return (
-        <View>
-            <Text>Profile</Text>
-        </View>
-    )
-}
-
-export default ProfileScreen
