--- src/theme/index.js
@@ -1,3 +1,0 @@
-export const theme = {
-
-};
